# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Salón de Belleza",
        "description": "API REST para gestión de citas y servicios del salón de belleza",
        "contact": {
            "name": "Equipo de Desarrollo",
            "email": "admin@salon.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor API"
        }
    ],
    "paths": {
...
        "/api/servicios/{id}": {
            "get": {
                "tags": [
                    "Servicios"
                ],
                "summary": "Obtener un servicio específico",
                "description": "Obtiene los detalles de un servicio por su ID",
                "operationId": "getServicio",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del servicio",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Servicio encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Servicio encontrado"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nombre": {
                                                    "type": "string",
                                                    "example": "Corte de Cabello"
                                                },
                                                "precio": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 80
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Servicio no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Servicio no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}